# app/api/snippets/restore/route.ts
import { dbForRequest, rowToSnippet } from "@/lib/db";
import { randomUUID } from "node:crypto";
import { NextResponse } from "next/server";
import { LANGUAGES } from "@/lib/languages";
import {
  sanitizeTags,
  sanitizeModel,
  sanitizeKind,
  sanitizeColor,
  validTimestampOr,
} from "@/lib/api-utils";

const validLanguages = new Set<string>(LANGUAGES.map((l) => l.value));

// Re-insert a previously deleted snippet, preserving its fields (favorite,
// model, usage counts, timestamps). Backs undo-after-delete. The restored row
// gets a new autoincrement id. Every field is validated/normalized the same way
// the create route does, plus a few restore-specific guards, so a crafted
// request can't inject a fractional copy_count (which would later break reads),
// a "truthy" favorite, or an out-of-range timestamp.
export async function POST(request: Request) {
  const db = dbForRequest(request);
  try {
    const body = await request.json();
    const {
      uuid,
      title,
      description,
      code,
      language,
      tags,
      favorite,
      model,
      kind,
      color,
      template,
      copy_count,
      last_used_at,
      created_at,
      updated_at,
    } = body;

    // Fast path for undo-after-delete: the row is only soft-deleted, so clear
    // its tombstone in place. This keeps the same uuid, so the undelete syncs
    // as an ordinary update rather than creating a duplicate on other machines.
    if (typeof uuid === "string" && uuid) {
      const undelete = db
        .prepare(
          "UPDATE snippets SET deleted = 0, updated_at = datetime('now') WHERE uuid = ?"
        )
        .run(uuid);
      if (undelete.changes > 0) {
        const row = db
          .prepare("SELECT * FROM snippets WHERE uuid = ?")
          .get(uuid) as Record<string, unknown>;
        return NextResponse.json(rowToSnippet(row), { status: 201 });
      }
    }

    if (
      typeof title !== "string" ||
      !title ||
      typeof code !== "string" ||
      !code ||
      typeof language !== "string" ||
      !validLanguages.has(language)
    ) {
      return NextResponse.json(
        { error: "A valid snippet is required" },
        { status: 400 }
      );
    }

    if (title.length > 255) {
      return NextResponse.json(
        { error: "Title must be 255 characters or fewer" },
        { status: 400 }
      );
    }

    const sanitizedTags = sanitizeTags(tags);
    const sanitizedModel = sanitizeModel(model);
    const sanitizedKind = sanitizeKind(kind);
    const sanitizedColor = sanitizeColor(color);

    // Must be a non-negative integer: a REAL (e.g. 1.5) stored in the INTEGER
    // column makes the desktop's typed read fail for the whole list.
    const sanitizedCopyCount =
      typeof copy_count === "number" && Number.isFinite(copy_count)
        ? Math.max(0, Math.floor(copy_count))
        : 0;

    const now = new Date().toISOString().slice(0, 19).replace("T", " ");
    const createdAt = validTimestampOr(created_at, now);
    const updatedAt = validTimestampOr(updated_at, now);
    const lastUsedAt = validTimestampOr(last_used_at, null);

    const stmt = db.prepare(`
      INSERT INTO snippets (uuid, title, description, code, language, tags, favorite, model, kind, color, template, copy_count, last_used_at, created_at, updated_at)
      VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
    `);

    const result = stmt.run(
      typeof uuid === "string" && uuid ? uuid : randomUUID(),
      title,
      typeof description === "string" ? description : "",
      code,
      language,
      JSON.stringify(sanitizedTags),
      favorite === true ? 1 : 0,
      sanitizedModel,
      sanitizedKind,
      sanitizedColor,
      template === true ? 1 : 0,
      sanitizedCopyCount,
      lastUsedAt,
      createdAt,
      updatedAt
    );

    const restored = db
      .prepare("SELECT * FROM snippets WHERE id = ?")
      .get(result.lastInsertRowid) as Record<string, unknown>;

    return NextResponse.json(rowToSnippet(restored), { status: 201 });
  } catch (error) {
    console.error("Failed to restore snippet:", error);
    return NextResponse.json(
      { error: "Failed to restore snippet" },
      { status: 500 }
    );
  }
}